requirements.txt
Flask==1.1.2
Pillow==7.0.0
torch==1.6.0
torchvision==0.7.0